--- src/modules/actions/index.jsx
@@ -1,258 +1,0 @@
-import React, { forwardRef } from 'react'
-
-import { getQualification } from 'cozy-client/dist/models/document'
-import { getBoundT } from 'cozy-client/dist/models/document/locales'
-import { isFile } from 'cozy-client/dist/models/file'
-import ActionsMenuItem from 'cozy-ui/transpiled/react/ActionsMenu/ActionsMenuItem'
-import Divider from 'cozy-ui/transpiled/react/Divider'
-import Icon from 'cozy-ui/transpiled/react/Icon'
-import DownloadIcon from 'cozy-ui/transpiled/react/Icons/Download'
-import HistoryIcon from 'cozy-ui/transpiled/react/Icons/History'
-import UnqualifyIcon from 'cozy-ui/transpiled/react/Icons/LabelOutlined'
-import QualifyIcon from 'cozy-ui/transpiled/react/Icons/Qualify'
-import RenameIcon from 'cozy-ui/transpiled/react/Icons/Rename'
-import RestoreIcon from 'cozy-ui/transpiled/react/Icons/Restore'
-import TrashIcon from 'cozy-ui/transpiled/react/Icons/Trash'
-import ListItemIcon from 'cozy-ui/transpiled/react/ListItemIcon'
-import ListItemText from 'cozy-ui/transpiled/react/ListItemText'
-
-import { downloadFiles, restoreFiles } from './utils'
-import { HOME_LINK_HREF } from 'constants/config'
-import { isEncryptedFolder, isEncryptedFile } from 'lib/encryption'
-import { navigateToModal } from 'modules/actions/helpers'
-import DeleteConfirm from 'modules/drive/DeleteConfirm'
-import { startRenamingAsync } from 'modules/drive/rename'
-import { openExternalLink as openExtLink } from 'modules/public/helpers'
-
-export { share } from './share'
-
-export const download = ({ client, t, vaultClient, showAlert }) => {
-  const label = t('SelectionBar.download')
-  const icon = DownloadIcon
-
-  return {
-    name: 'download',
-    label,
-    icon,
-    displayCondition: files => {
-      // We cannot generate archive for encrypted files, for now.
-      // Then, we do not display the download button when the selection
-      // includes an encrypted folder or several encrypted files
-      return (
-        files.length > 0 &&
-        !files.some(file => isEncryptedFolder(file)) &&
-        !(files.length > 1 && files.some(file => isEncryptedFile(file)))
-      )
-    },
-    action: files =>
-      downloadFiles(client, files, { vaultClient, showAlert, t }),
-    Component: forwardRef(function Download(props, ref) {
-      return (
-        <ActionsMenuItem {...props} ref={ref}>
-          <ListItemIcon>
-            <Icon icon={icon} />
-          </ListItemIcon>
-          <ListItemText primary={label} />
-        </ActionsMenuItem>
-      )
-    })
-  }
-}
-
-export const hr = () => {
-  return {
-    name: 'hr',
-    icon: 'hr',
-    displayInSelectionBar: false,
-    Component: forwardRef(function hr(props, ref) {
-      return <Divider ref={ref} className="u-mv-half" />
-    })
-  }
-}
-
-export const trash = ({
-  t,
-  pushModal,
-  popModal,
-  hasWriteAccess,
-  refresh,
-  byDocId,
-  isOwner
-}) => {
-  const icon = TrashIcon
-
-  return {
-    name: 'trash',
-    icon,
-    displayCondition: files => files.length > 0 && hasWriteAccess,
-    action: files =>
-      pushModal(
-        <DeleteConfirm
-          files={files}
-          afterConfirmation={() => {
-            refresh()
-          }}
-          onClose={popModal}
-        />
-      ),
-    Component: forwardRef(function Trash(props, ref) {
-      const sharedWithMe =
-        byDocId !== undefined &&
-        byDocId[props.docs[0].id] &&
-        !isOwner(props.docs[0].id)
-      const label = sharedWithMe ? t('toolbar.leave') : t('SelectionBar.trash')
-
-      return (
-        <ActionsMenuItem {...props} ref={ref}>
-          <ListItemIcon>
-            <Icon icon={icon} color="var(--errorColor)" />
-          </ListItemIcon>
-          <ListItemText
-            primary={label}
-            primaryTypographyProps={{ color: 'error' }}
-          />
-        </ActionsMenuItem>
-      )
-    })
-  }
-}
-
-export const rename = ({ t, hasWriteAccess, dispatch }) => {
-  const label = t('SelectionBar.rename')
-  const icon = RenameIcon
-
-  return {
-    name: 'rename',
-    label,
-    icon,
-    displayCondition: selection => hasWriteAccess && selection.length === 1,
-    action: files => dispatch(startRenamingAsync(files[0])),
-    Component: forwardRef(function Rename(props, ref) {
-      return (
-        <ActionsMenuItem {...props} ref={ref}>
-          <ListItemIcon>
-            <Icon icon={icon} />
-          </ListItemIcon>
-          <ListItemText primary={label} />
-        </ActionsMenuItem>
-      )
-    })
-  }
-}
-
-export const qualify = ({ t, lang, navigate, pathname }) => {
-  const label = t('SelectionBar.qualify')
-  const scannerT = getBoundT(lang || 'en')
-
-  return {
-    name: 'qualify',
-    label,
-    icon: QualifyIcon,
-    displayCondition: selection =>
-      selection.length === 1 && isFile(selection[0]),
-    action: files =>
-      navigateToModal({ navigate, pathname, files, path: 'qualify' }),
-    Component: forwardRef(function Qualify(props, ref) {
-      const file = props.docs[0]
-      const fileQualif = getQualification(file)
-
-      return (
-        <ActionsMenuItem {...props} ref={ref}>
-          <ListItemIcon>
-            <Icon icon={fileQualif ? QualifyIcon : UnqualifyIcon} />
-          </ListItemIcon>
-          <ListItemText primary={fileQualif ? t('Scan.requalify') : label} />
-          {fileQualif && (
-            <ListItemText
-              secondary={scannerT(`Scan.items.${fileQualif.label}`)}
-              secondaryTypographyProps={{ variant: 'caption' }}
-              className="u-ta-right"
-            />
-          )}
-        </ActionsMenuItem>
-      )
-    })
-  }
-}
-
-export const versions = ({ t, navigate, pathname }) => {
-  const label = t('SelectionBar.history')
-  const icon = HistoryIcon
-
-  return {
-    name: 'history',
-    label,
-    icon,
-    displayCondition: selection =>
-      selection.length === 1 && isFile(selection[0]),
-    action: files =>
-      navigateToModal({ navigate, pathname, files, path: 'revision' }),
-    Component: forwardRef(function History(props, ref) {
-      return (
-        <ActionsMenuItem {...props} ref={ref}>
-          <ListItemIcon>
-            <Icon icon={icon} />
-          </ListItemIcon>
-          <ListItemText primary={label} />
-        </ActionsMenuItem>
-      )
-    })
-  }
-}
-
-export const restore = ({ t, refresh, client }) => {
-  const label = t('SelectionBar.restore')
-  const icon = RestoreIcon
-
-  return {
-    name: 'restore',
-    label,
-    icon,
-    action: async files => {
-      await restoreFiles(client, files)
-      refresh()
-    },
-    Component: forwardRef(function Restore(props, ref) {
-      return (
-        <ActionsMenuItem {...props} ref={ref}>
-          <ListItemIcon>
-            <Icon icon={icon} />
-          </ListItemIcon>
-          <ListItemText primary={label} />
-        </ActionsMenuItem>
-      )
-    })
-  }
-}
-
-export const openExternalLink = ({ t, isSharingShortcutCreated, link }) => {
-  const label =
-    link === HOME_LINK_HREF
-      ? t('Share.create-cozy')
-      : isSharingShortcutCreated
-      ? t('toolbar.menu_sync_cozy')
-      : t('toolbar.add_to_mine')
-  const icon =
-    !isSharingShortcutCreated || link === HOME_LINK_HREF
-      ? 'to-the-cloud'
-      : 'sync'
-
-  return {
-    name: 'openExternalLink',
-    label,
-    icon,
-    action: () => {
-      openExtLink(link)
-    },
-    Component: forwardRef(function OpenExternalLink(props, ref) {
-      return (
-        <ActionsMenuItem {...props} ref={ref}>
-          <ListItemIcon>
-            <Icon icon={icon} />
-          </ListItemIcon>
-          <ListItemText primary={label} />
-        </ActionsMenuItem>
-      )
-    })
-  }
-}
