+++ src/modules/actions/rename.jsx
@@ -1,0 +1,39 @@
+import React, { forwardRef } from 'react'
+
+import ActionsMenuItem from 'cozy-ui/transpiled/react/ActionsMenu/ActionsMenuItem'
+import Icon from 'cozy-ui/transpiled/react/Icon'
+import RenameIcon from 'cozy-ui/transpiled/react/Icons/Rename'
+import ListItemIcon from 'cozy-ui/transpiled/react/ListItemIcon'
+import ListItemText from 'cozy-ui/transpiled/react/ListItemText'
+
+import { startRenamingAsync } from 'modules/drive/rename'
+
+const makeComponent = (label, icon) => {
+  const Component = forwardRef((props, ref) => {
+    return (
+      <ActionsMenuItem {...props} ref={ref}>
+        <ListItemIcon>
+          <Icon icon={icon} />
+        </ListItemIcon>
+        <ListItemText primary={label} />
+      </ActionsMenuItem>
+    )
+  })
+  Component.displayName = 'Rename'
+
+  return Component
+}
+
+export const rename = ({ t, hasWriteAccess, dispatch }) => {
+  const label = t('SelectionBar.rename')
+  const icon = RenameIcon
+
+  return {
+    name: 'rename',
+    label,
+    icon,
+    displayCondition: selection => hasWriteAccess && selection.length === 1,
+    action: files => dispatch(startRenamingAsync(files[0])),
+    Component: makeComponent(label, icon)
+  }
+}
